+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **Changed**: `groups` to `layers`
 - **Changed**: groups and edit, merged together
 - **Changed**: groups modal to sidebar
+- **Changed**: `LayerItem` design
 - **Changed**: edit menu
 
 ## 0.5.6 : 2024-07-18